[frontport] Extract CommonCliOptions from the linera binary into linera-service library (#5529)
